fix potential issues and security patches

+++ lib/actions/DeleteAnswer.ts
@@ -10,6 +10,7 @@ import Question from "@/database/question.model";
 import Vote from "@/database/vote.model";
 import { revalidatePath } from "next/cache";
 import ROUTES from "@/routes";
+import mongoose from "mongoose";
 
 const deleteAnswer = async (params: { answerId: string }) => {
   await dbConnect();
@@ -17,30 +18,44 @@ const deleteAnswer = async (params: { answerId: string }) => {
   const { answerId } = validatedData.data;
   const auth_session = await auth();
   const user = auth_session?.user;
+
+  if (!user) {
+    throw new Error("Not authenticated");
+  }
+  const session = await mongoose.startSession();
+
   try {
+    session.startTransaction();
     const answer = await Answer.findById(answerId);
     if (!answer) throw new Error("Answer not found");
+
+    if (answer.author.toString() !== user.id) {
+      throw new Error("Not authorized");
+    }
 
     await Question.findByIdAndUpdate(
       answer.question,
       {
         $inc: { answers: -1 },
       },
-      { new: true }
+      { new: true, session }
     );
 
     await Vote.deleteMany({
       actionId: answerId,
       actionType: "answer",
-    });
+    }, { session });
 
-    await Answer.findByIdAndDelete(answerId);
+    await Answer.findByIdAndDelete(answerId).session(session);
 
     revalidatePath(ROUTES.PROFILE(user?.id as string));
 
     return { success: true };
   } catch (e) {
+    await session.abortTransaction();
     return actionError(e);
+  } finally {
+    session.endSession();
   }
 };
